move typefixture to mypy/test; update imports
