+++ mypy/test/testinfer.py
@@ -6,7 +6,7 @@ from mypy.myunit import Suite, assert_equal, assert_true
 from mypy.checkexpr import map_actuals_to_formals
 from mypy.nodes import ARG_POS, ARG_OPT, ARG_STAR, ARG_STAR2, ARG_NAMED
 from mypy.types import AnyType, TupleType, Type, TypeOfAny
-from mypy.typefixture import TypeFixture
+from mypy.test.typefixture import TypeFixture
 
 
 class MapActualsToFormalsSuite(Suite):
